chore: cleanup proxy files (#10268)
- merge `readonly.js` into `proxy.js` and get rid of sub folder
- extract types into `d.ts` file and properly document the properties
- type tweaks
