--- packages/svelte/src/internal/client/private.d.ts
@@ -1,8 +1,0 @@
-export type TaskCallback = (now: number) => boolean | void;
-
-export type TaskEntry = { c: TaskCallback; f: () => void };
-
-export interface Task {
-	abort(): void;
-	promise: Promise<void>;
-}
