+++ packages/svelte/src/internal/client/types.d.ts
@@ -10,6 +10,7 @@ import {
 	DYNAMIC_ELEMENT_BLOCK,
 	SNIPPET_BLOCK
 } from './block.js';
+import type { READONLY_SYMBOL, STATE_SYMBOL } from './proxy.js';
 import { DERIVED, EFFECT, RENDER_EFFECT, SOURCE, PRE_EFFECT, LAZY_PROPERTY } from './runtime.js';
 
 // Put all internal types in this file. Once we convert to JSDoc, we can make this a d.ts file
@@ -387,4 +388,34 @@ export type Render = {
 export type Raf = {
 	tick: (callback: (time: DOMHighResTimeStamp) => void) => any;
 	now: () => number;
+};
+
+export interface Task {
+	abort(): void;
+	promise: Promise<void>;
+}
+
+export type TaskCallback = (now: number) => boolean | void;
+
+export type TaskEntry = { c: TaskCallback; f: () => void };
+
+export interface ProxyMetadata<T = Record<string | symbol, any>> {
+	/** A map of signals associated to the properties that are reactive */
+	s: Map<string | symbol, SourceSignal<any>>;
+	/** A version counter, used within the proxy to signal changes in places where there's no other way to signal an update */
+	v: SourceSignal<number>;
+	/** `true` if the proxified object is an array */
+	a: boolean;
+	/** Immutable: Whether to use a source or mutable source under the hood */
+	i: boolean;
+	/** The associated proxy */
+	p: ProxyStateObject<T> | ProxyReadonlyObject<T>;
+}
+
+export type ProxyStateObject<T = Record<string | symbol, any>> = T & {
+	[STATE_SYMBOL]: ProxyMetadata;
+};
+
+export type ProxyReadonlyObject<T = Record<string | symbol, any>> = ProxyStateObject<T> & {
+	[READONLY_SYMBOL]: ProxyMetadata;
 };
