+++ packages/svelte/src/internal/index.js
@@ -45,7 +45,7 @@ export * from './client/each.js';
 export * from './client/render.js';
 export * from './client/validate.js';
 export { raf } from './client/timing.js';
-export { proxy, readonly, unstate } from './client/proxy/proxy.js';
+export { proxy, readonly, unstate } from './client/proxy.js';
 
 export { create_custom_element } from './client/custom-element.js';
