+++ packages/svelte/src/motion/tweened.js
@@ -81,7 +81,7 @@ function get_interpolator(a, b) {
  */
 export function tweened(value, defaults = {}) {
 	const store = writable(value);
-	/** @type {import('../internal/client/private').Task} */
+	/** @type {import('../internal/client/types').Task} */
 	let task;
 	let target_value = value;
 	/**
@@ -95,7 +95,7 @@ export function tweened(value, defaults = {}) {
 		}
 		target_value = new_value;
 
-		/** @type {import('../internal/client/private').Task | null} */
+		/** @type {import('../internal/client/types').Task | null} */
 		let previous_task = task;
 
 		let started = false;
